extracted API logic from component
